[Requirement #11864] - DICOM folders should support DICOM query/retrieve
Fixed a kind of bug with the DICOM filestream functionality that was causing
some images received by ImageDrive to not be associated with the proper
association.

The problem was caused by DICOM files that had a large private tag
encoded in them before the "stop tag" that we wanted to parse to.  The
data was being passed to the callback function before we had a full header
that we could decide what to do with the file.  Luckily the SOP Instance UID
was parsed, so we at least knew that.  In any case, added code to buffer
data received in a MemoryStream until we pass it off to actual stream handler
with a properly populated header so that we can decide where to store the
file.

--HG--
branch : Rockstar