[Ticket #11958] Alerts due to bad configured hostname gives cryptic error message
Added a try/catch around the DNS call to resolve the hostname for clients
connecting to a server.  Also added in a more descriptive error message in
the FailureDescription for the SCU that when this exception occurs, it is
an error when connecting to a specific AE title.

--HG--
branch : Rockstar