[Ticket #11966] DICOM send operation is aborted if one instance has an unsupported transfer syntax
A failure condition when a matching presentation context was not found was not
being handled properly.  An output DicomMessage was not set when a matching
presentation context is not found, and we referencing the null message, which
in turn was causing a null reference exception.  Changed for logging/error message
purposes how we get info on the sop that failed to send.

--HG--
branch : Rockstar