[Ticket #11821] - Update database schema upgrade scripts [Ticket #12247] - ImageServer will support storing orders in its database
Added ServerPartitionGUID into the ProcedureCode + Staff tables, they're needed
to associate the data with specific partitions and avoid conflicts.  Also reworked
the order status to not be so HL7 centric, and changed also how the order status
is sent over the create order.  Also, implemented the actual populating of
the order, and updating if status updates.

--HG--
branch : Orion