[Requirement #12217] - Cleome RIS must be localization-ready, with a test Spanish translation
There seems to be a race-condition with loading the localizations, where it is possible that the SR variable is not yet populated with the strings when the body onload event fires. When this happens, a "script error" dialog box was being shown to the user. Given that this happens *very* infrequently, it is easier to mitigate the impact rather than try to resolve the underlying cause of the race. In order to limit the impact of potential script errors on the end-user:
 * added code to DHtmlComponentControl to suppress script dialogs when not running in DEBUG mode.
 * added a window.onerror handler, and some infrastructure to route javascript error messages to the Workstation log file; this will be necessary to troubleshoot issues in production builds where script error dialogs are suppressed.

--HG--
branch : Orion