[Requirement #12217] - Cleome RIS must be localization-ready, with a test Spanish translation
Re-ordered the code so that the SR script node is added to the DOM before setting the src property, as this is apparently a best practice (according to bloggers everywhere). It does seem to reduce the incidence of the previously reported race condition (in fact, since making this change I have yet to see the race condition manifest at all).

--HG--
branch : Orion