[Ticket #12253] - IntegrationServer wil process ORM^O01 messages and transmit them to the ImageServer
Cleaned up a number of issues with the HL7 ORM processing.  The IntegrationServer
was not dealing with missing fields that well, made some adjustments to how I
grabbed fields out of the message to get that working better.  Cleaned up
some compilation issues in the prior commit.  Also, fixed an issue with the
PatientClass column not being long enough.  Also, had to make some updates to
the ImageServer's ORM to handle the fact that the Order table is a keyword,
some of the formulated SQL didn't work because of this.

--HG--
branch : Orion