[Ticket #12249] - ImageServer will have administrative GUI to display orders
Added additional columns displayed on the order page, including the
Requested Procedure, reason for study,  priority, and Insert time. Also changed
the scheduled time so its displayed as a DateTime instead of a Date.   Also fixed
an issue with the search not working for status values.  This should be resolved
now.

--HG--
branch : Orion