[Requirement #12263] - The Studies page will allow searching for studies by QC Status
Prior checkin of the ugprade script didn't include the QCStatusEnum values being
inserted.

Also gave a shot at properly inserting the QCStatusEnum values according to
existing status values in the QCOutput column,and then changing the column
so it doesn't allow nulls.  Note that I changed the default creation script
for the Study table to not allow nulls for QCStatusEnum

--HG--
branch : Orion