[Ticket #12250] - #12250 (Orion) - ImageServer studies will display linked order in study details, if it exists
Added new details tab with Order information.  Decided to leave this visible for
the time being for all users.  When there is not a linked order, the tab is just
empty.  The tab has all the available fields from the order displayed, when an order
is linked to the study.

Also made some updates to the WebServices test app to send the CreateOrUpdateOrder
request.

Finally, made the actual update to link the Order to the Study in the CreateOrUpdateOrderProcessor.

--HG--
branch : Orion