[Ticket #12251] - ImageServer will link studies to orders based on accession number when processing studies.
Added extension to populate the OrderKey in the studies table when a new study
is coming in that has a matching order (bassed on accession number).  Required
some updates to the IInsertExtension interface for adding extensions when inserting
a new study in.  The partitionKey had to be passed in to search for the matching
order.

--HG--
branch : Orion