[Ticket #12249] - ImageServer will have administrative GUI to display orders
Found an issue with deleting an Order, it was not cleaning up/removing the patient
if the patient didn't have a referenced study.  This is now cleaned up.  Required
some changes to the entity broker, subselects were not working with the Order
table, and added some subselect criteria for the patient broker so we
could only delete when no order/study was linked to the patient.  Also had to
make some changes to the delete dialog to pass around the patientguid.

--HG--
branch : Orion