[Ticket #12247] - ImageServer will support storing orders in its database. [Ticket #12256] - Patient Update and Merge Messages received through the ImageServer API will update orders in the database
Decided to denormalize the Orders table, and store the Patient Id/Issuer Of Patient Id/PatientsName
all in the Orders table.  (I will remove the link entirely to the Patient table
in the next commit).  This update includes adding the fields into the database
and adding appropriate indexes for querying.

Also implemented Patient Updates/Merges with orders.  We now will will search
for matching orders on the update/merge and update the Name/Id/Issuer of Id as
appropriate in the database, according to the update and merge messages.  This
functionality is working.

Finally, added into the web services test tool the capability to send update
and merge messages for studies to test the api.

--HG--
branch : Orion