[Ticket #12247] - ImageServer will support storing orders in its database. [Ticket #12249] - ImageServer will have administrative GUI to display orders
Removed the PatientGUID column from the Order table.  We now no longer maintain
a reference to the Patient for orders, the patient related info is just stored
with the order.  This required backing out a bunch of changes that tried to
deal with order references to the Patient table.  This should be a cleaner
design, and we shouldn't have to deal with Patient ID issues with linking as
much.  There was no real usage of the order to patient link in this release,
besides attempting to maintain the link, so it probably makes sense for now
not to have the link.

--HG--
branch : Orion