[Requirement #12342] - Orders in the database will be purged if no matching study arrives within a configurable time period
Reworked the code for adding partitions so that it calls a web service to do
the adding.  The web service is then processed in the ShredHostService.  This
allows a new ServerPartitionAddedArgs event in the ImageServer that I made a
plugin for that runs every time a partition is created.  This makes it so the
new ServicelockTypeEnum.PartitionOrderPurge is added when a new partition is
added.  With the combination of the constructor and the event, all the partitions
should have a ServiceLockTypeEnum added on upgrade or new partitions.

Also fixed a small bug in the ExternalRequestQueue page due to a change in the
name of a resource string.

--HG--
branch : Orion