[Requirement #12257] - Studies can be auto-routed after passing QC tests
Introduced the new auto-route action for doing study level routing.  The route
rules are applied at the study level, and a single workitem is created to
auto-route the study.  it is also possible to specify a QC status, and a delay
in minutes to wait to perform the auto-route.  If addendum images are applied to
the study, or if more than one auto-route rule applies tot he same destination
device, the same WorkQueue entry will be reused, and pushed to the new time.

This required a new ServerRuleTypeEnum, made the appropriate updates to handle
this including the database upgrade script.

--HG--
branch : Orion