[Ticket #12249] - ImageServer will have administrative GUI to display orders
Added the ability to search by Referring Physician's name into the GUI.
Required an exists clause for querying and then the field to be included.
Note that queries for first/last name can be done one of two ways:

Last*^First*, or Last*,First*

Wildcards are automatically added to the fields.

--HG--
branch : Orion