[Ticket #12251] - ImageServer will link studies to orders based on accession number when processing studies.
Added code to link the studies to orders after an edit of the study.  Realized
this wasn't implemented properly when reviewing TRs post sprint.  This should now
be working properly.  Also updated the order to study link so that when an order
comes in, we search for more than one study that may be linked to the study, as
matched in the TR.  The initial implementation only took the first study found
when searching for a match.

--HG--
branch : Orion