[Ticket #12206] - N Cleome ImageServers in a cluster will process work items Nx faster than one ImageServer
Checked all the places where we were doing File.Exists() checks in the code
and tried to eliminate them.  Many of the places were before a call to File.Delete
or FileUtils.Delete(), which both fail silently if the file does not exist,
so these were all not necessary.  Also changed some File.Delete() calls
within the ImageServer source code to use FileUtils.Delete instead, this seemed
to make sense to be consistent in usage.

Hopefully this will help with performance when using the NAS.

--HG--
branch : Orion