[Ticket #12251] - ImageServer will link studies to orders based on accession number when processing studies.
Updated the code, as per the risk review, to match on both Accession Number and patient ID when
matching a study to an order.  Also updated the order on a patient Merge so that
we update the order before inserting an entry in the workqueue for updating the
studies, just to ensure the orders are updated first.

--HG--
branch : Orion