[Ticket #12380] - Optimize query-related code in the ImageServer and DICOM toolkit
Prior optimization for reading in the settings values for validating
Vr Lengths/Values/Ignoring Out of range tags was not done correctly.  I just
changed a property value to static, which would cause the value to be updated
overtime. by different instancfes of the DIcomAttributeCollection.

Changed so that the default values are loaded once and stored in a static, and then
set in the constructors from the static values each time the class instance is
created.

--HG--
branch : Orion