[Requirement #12337] - Rules can be configured to identify orders intended for QC
Initial implementation of the functionality, waiting on some further feedback from
customer as to if this makes sense.  Added a new setting to ExternalRequestSettings,
IntendedForQCProcedures, which is a backslash delimited list of procedure codes
that identify a study as being intended for QC.  Also made the setting be stored
in the enterprise settings store so it can be updated in the workstation GUI.  The
code currently looks at the RequestedProcedureCode field in the order to look for
matches.  The match is based on a "contains" check.  If value is included in the
text of the procedure code, it is considered that the order is intended for QC.

Modified the Orders table to include a QCExpected column that is a bit, cannot
be null, and defaults to 0.

Also added into the orders page in the GUI a display if the order is inteded for
QC, and the ability to search based on if the order is intended for QC via
a drop down.

Note that the field searched for matches might be updated, and still have to review
the labeling on the page to make sure everything is good.

--HG--
branch : Orion