[Ticket #12208] - Cleome ImageServers in a cluster will not exceed device throttling configuration while auto-routing or sending studies
The prior code was doing a sort of the curently active requests by scheduled
date and allowing an auto-route/web move to continue if it was one of the
scheduled entries in order that was less than the Threshhold count.  If for
some reason a later scheduled entry was active and had gotten past the CanStart
check, it could have caused more than the threshhold count to run at the same
time.

Also found a small bug that I cleaned up related to the WebMove processor
checking for a reconcile status, and not delaying the WorkItem properly in
its CanStart() check.

--HG--
branch : Orion