[Ticket #12380] - Optimize query-related code in the ImageServer
Reworked the way the StudyXml code is interacted with to add flexibility with
how the Study XML is created so that we may be able to move away from an
XmlDocument based Memento in the future, to improve performance.  This also
would allow cached XML fragments to be passed through in the memento when
repeatedly writing out the file.

--HG--
branch : Orion