[Ticket #12340] - ImageServer will email key people a configurable amount of time after the patient has arrived
Introduced support int he ExternalRequestQueue to schedule an entry to a sepcific
time.  This is needed for scheduling the mail request.  I had to do database
updates to make it work.  Note that the upgrade scripts checked in can just be
used to update an existing database.  Hopefully this should work.

--HG--
branch : Orion