[Ticket #12206] - N Cleome ImageServers in a cluster will process work items Nx faster than one ImageServer
Reworked the actual StudyXml output.  We now actually cache the Xml text fragment of
data to be written, so saving the file just requires reassembling all the
fragments and flushing them to disk.  This should be more efficient than what
was done before, and may be able to go even further optimizing it.

a preliminary test of processing a 2400 slice MR shows around a 20% improvement
in the time to save the Study XML.  Need to do some further comparisons with
processing times on larger studies.  Also need to run the new code through the
profiler.  Note that the new code can be shut off by changing the
StudyXmlOutputSettings.OptimizedMemento parater to false.

--HG--
branch : Orion