[Ticket #12373] - ImageServer attempts compression on MR Spect non-images and fails
Updated the DicomMessageBase.ChangeTransferSyntax method to take into account
when MR spectroscopy data is attempting to be compressed, we just let the change
of transfer syntax go through, instead of throwing an exception.

Also updated the sample rules to include a test to exclude MR spectroscopy objects
from being compressed.  This secondary change is not necessary, since we don't
really include exceptions for every non-image object, but thought it necessary
to include it for clarity.

--HG--
branch : Orion