[Ticket #11587] - ImageServer cannot compress a PT study from a customer (not sure why)
Made changes to make it so SOP Compress actions do not cause processing failure
if they fail due to an exception.  We just skip over the compression and leave
the file "as is" without compressing it, logging a warning message.

Note that in this specific case, the compression was failing because of the signed
pixel data, and the fact that the PET object does not support a modality LUT
(a note in the IOD limits the rescale intercept to 0) which would allow us to
change the pixel data to unsigned.  So, there's no way to JPEG lossless compress
these specific SOPs.

--HG--
branch : Orion