[Ticket #12395] - DICOM receive performance degrades over time
One final small optimization to the study creation.  Added in caching of the
xml fragment of the base instances, so that they don't have to be recreated.

Also made a final change to use the LargeMemoryStream instead of the MemoryStream
to store the StudyXml in memory before its written out.  I tried a couple other
things, such as using a BufferedStream and a custom stream that would write to
both the gzip + the file at the same time instead of a memory stream.  In the
end, writing the file out all at once seems to be the desired method based
on the profile.  Decided to go to the LargeMemoryStream (even though its slightly
less efficient) because the profile of the MemoryStream showed a fair amount of
time expanding the MemoryStream + to help with memory usage.

--HG--
branch : Orion