[Ticket #12395] - DICOM receive performance degrades over time
Updated the default WorkQueueTypeProperties in the ImageServer.  The delay between
processing "batches" of images in a study has been dropped by default from 10
to 5 seconds.  The time after no images are received until the WorkQueue entry
completes has been dropped from 120 to 30 seconds.  Note that there was some
concern about this setting in cases when large (CR/DX/MG) images are transferred
over a slow network connection, so this was NOT dropped as low as the settings
recommended for Bridge/Synaptive.  Finally, changed the failure delay seconds
(the length a StudyProcess WorkItem is delayed when any kind of failure occurs)
from 180 seconds down to 60 seconds.  This also isn't as aggressive as the recommend
5 seconds, but its at least faster.

Note that these values will only be updated on upgrade if the original values
for the settings are the current setting values in the database.  If they
have been updated, they will stay at the updated value.

--HG--
branch : Orion