[Requirement #12340] - ImageServer will email key people a configurable amount of time after the patient has arrived
I mistakenly remove dthe constraint on the Revision column in a prior commit.  Jasper
added it back in, but did not give the constraint a name, which exists in the upgrade
script.  made  achange to add a name to the constraint so that the upgrade and fresh
database creation matches.

--HG--
branch : Orion