[Ticket #12419] - It is extremely difficult to configure ImageServer web services to use certificate
Removed the RemoteImageServerServiceProvider class and associated settings
because the ImageServer was no longer using the service provider, with the
exception of the open source ImageServer, which can in turn be moved to the
EnterpriseImageServerServiceProvider.  Updated the configuration files to
reflect this removal.

Also moved some configuration out of the WebServicesSettings into the
EnterpriseImageServerServiceSettings related to setting the security mode for
the WCF communication. The WebServicesSettings now just sets the base URL for
both the REST + WCF server.  It should not be needed for anything else.  This
change should make it so that the security mode for the messaging should be set
in one place in the EnterpriseServer settings store, and it should update both
the client and server's settings.

--HG--
branch : Orion