[Ticket #12479] - Web Components unable to display QCOrderNotificationRequest in External Request Queue
Updated the location of the request for QC notification into the WorkExternalRequest
source file.  Also had to move some of the resource strings around because of this.

While testing, found a few issues to clean up ont he Web GUI page for the
external request queue.  There was a Javascript error that was causing an
exception in IE that was only visible when debugging.  Also added in the
ScheduledTime to the page, since the QC order notification request uses this
and it wasn't displayed, so it wasn't clear why the entry wasn't being processed.

--HG--
branch : Orion