[Ticket #12486] - Orders admin search fields are inconsistent with the Studies pages
Checked the other pages, and the "search to" field has to add 24 hours to it to
get the search to work properly.  Added this in, and it appears to work.

Note that nothing was done yet to address the referring physicians searches.  Since
the database field is broken up into Given/Family/Middle name fields, the searching
like the other pages is more difficult.

--HG--
branch : Orion