[Ticket #12500] - Study Auto-Route is not working
The code for this never was finished/worked.  Not sure how I let this one
get through.  The standard AutoRoute WorkQueue type was being used and it didn't
work without all the SOP instances being inserted into the WorkQueueUid table.

Anyways, added a new WorkQueueTypeEnum, StudyAutoRoute.  It works almost identical
to the WebMoveStudy WorkQueue entry.  The WorkQueueUid table is not used, the
entry is just inserted and it runs.  The WorkQueueData column is used to
store a note that the queue entry completed, so that it can stay Idle and
then be removed after its complete.  This should work properly now.

--HG--
branch : Orion