[Ticket #12492] - ImageServer may fail to start after reboot
Discovered the UsageTracking Shred, StartupShred (for an Alert), ManifestCheckingShred,
and Remote WCF Services Shred all did not deal properly with the database not
being up.  Introduced a new class, ShredStartupHelper, to help test if the database
is operational for each of these services, and cleanly handle shutdown requests
when the database is down.

Also, updated some of the log messages so we don't have as much clutter in
the log file in this startup mode with the database down.

--HG--
branch : Orion