[Ticket #12510] - Study Details page shows QC Status even when not licensed
Reworked on the Study Details page the displaying of the QC status with
the patient panel.  Introduced a new extension point PatientSummaryPanelExtensionPoint
and implemented a plugin to setup the QC Status.  This allows the linking of the
status to the license.  I also cleaned up some of the QC fields being filled
in.  This resolves an issue that made it hard to synchronize the patient
panel with the open source build.

--HG--
branch : Orion