[Ticket #12520] - Unexpected exception warnings in the log when studies are tier-migrated
Surprised this wasn't cleaned up earlier.  The problem was that a cleanup of the
CopyDirectoryCommand.  The Dispose method was attempting to delete a backup
directory that in most cases wasn't set/needed and logging a needless warning
message.   Cleaned up the delete check to make sure that a directory has been
set before attempting to delete, and also cleaned up the logging messages a bit
to make them more helpful when we really do have a problem occurring with the
cleanup.

--HG--
branch : Orion