[Ticket #12520] - Unexpected exception warnings in the log when studies are tier-migrated
Addressed a few code comments.  Cleaned up the creation of directories to be
more efficient, and not make multiple attempts at the creation.

Also, deemed one code review as not relevant.  The TierMigrateItemProcessor
had a place where it issued multiple CreateDirectoryCommans.  This was done
so that rollback of just the created directories happens properly.  This could
be updated to create all the directories that are needed, but rollback would not
occur properly in this case.  Updated the code review comment.

--HG--
branch : Orion