[Ticket #12538] Access to data from outside the firewall should be through Web Components
Updated the DicomQueryApiController to use WCF services for performing Study
Root queries and Verification so the web app does not contact DICOM nodes
directly.  (Retrievals are already done by inserting a WorkItem into the queue,
 a rework was not required.)

Introduced some more service definitions into the ClearCanvas.Dicom.ServiceModel.Query
namespace where the configuration information for the query can be passed through
the service interface.  The existing IStudyRootQuery interface assumed the
service implementation already knew the application being queried when it was
created, so this didn't fit with how the services were needed.

Also introduced a new service definition for performing verifications (C-ECHO).
Both the Query + Verification services have baseline implementations in the
DICOM library, that can be used in applications directly.

Also, initially added a retrieve service definition.  Turned out this was not
needed, but left the definition for future use.

Note that this checkin modified some of the baseline query code used in the
Workstation, query support in the workstation should be retested after the
change.

With this change, the IIS app no longer needs to connect to DICOM applications
directly.

--HG--
branch : Polaris