[Ticket #12653] Audit Logging in ImageDrive
Added auditing of vfs copies, moves, deletes, and deidentification.  Also cleaned
some of the audit code for DICOM associations that was duplicated and moved
the assocation auditing class into the Research.Audit namespace.

The deidentification needs some work, because of the way its implemented,
we have the destination folder, but not the created study in the audit
message, would like to do this better.

--HG--
branch : Polaris