[Ticket #12538] Access to data from outside the firewall should be through Web Components
Still having some problems with the de-identifier not running correctly.
Made some updates to the DicomFile.Load() routine so that if a stream is not
seekable, and the bypass pixel data/store pixel references options are set,
we actually just read and throw out the data in the file, instead of trying
a Seek() that ends up throwing an exception.  Note that the StorePixelReferences
option cannot work in this case, but at least we can initially parse the file.

--HG--
branch : Polaris