[Ticket #12653] Audit Logging in ImageDrive
Updated audit logging so the client IP Address of the user who makes a request
is recorded in the audit log.  We stash the ipaddress in a new CreatorIpAddress
stored in the RequestBase class.  This is automatically filled in.  This field
is then passed into the appropriate audit log methods for use in the
AuditActiveParticipant objects.

Also, made some general code cleanup while I was at it.  Changed the
research audit helpers so they take a user participant in the constructors,
because these are always added.  Also made all of the audit helpers consistent
in adding the process participant into the audit message.  Finally, did a bit
of cleanup on how the AuditSource is logged after looking over some example
IHE logs.

Overall, the logs should be a bit more consistent and cleaner now.

--HG--
branch : Polaris