[Ticket #12538] Access to data from outside the firewall should be through Web Components
Some more cleanup to get the remote blob store to work better with the webstation.
Discovered that with large images, or else with PET/CT studies, the webstation
was freezing up.  Upon debugging, it was two things, I had put the timeout for
REST really high to handle large Uploads. I adjusted so there are now two timeouts
that are used, one for reading, the other for writing.  The previous timeout was
60 minutes, dropped the read timeout to 5 minutes for now by default.

Second, had to increase the maxconnection setting from 20 to 200 in the shredhost
config file.  The 20 value ended up being too small when there was a lot of
activity on the server, especially with prefetching from the webstation.

Discovered an area where we wre not disposing a Response object properly when
we open a read stream with the remote blob store.  Created a wrapper that will
dispose the response when the stream is disposed.

Finally, also discovered in this heavy load situation, I had to set
the KeepAlive setting to false and the ProtocolVersion to HttpVersion.Version10
when doing the get request to open a file on the remote blob store.  This seemed
to elminate some timeout/connection being dropped errors in some situations under
heavy load.

This seems to address the issue when I point my local webstation on my dev system
back to NightFury's remote blob store, which should be a worst case scenario.
Need to test further on the deployment system.

--HG--
branch : Polaris