[Ticket #12538] Access to data from outside the firewall should be through Web Components
I still saw some issues with the webstation timing out after viewing images for a while.
Noticed when doing "netstat -n", there was a large number of connections to the server
in the CLOSE_WAIT state.  Tracked this down to when reading the DicomFile, the Stream
was not having Close() called on it.

Also found some areas for cleanup in Close/Dispose handling in the REST client.
Need to test this in the production environment.  I disabled setting KeepAlive/
ProtocolVersion to 1.0 for the web request, because this seems to eliminate an
excessive amount of connections in TIME_WAIT state.

Need to do some more testing while using the webstation in the production
environment to make sure this works.

--HG--
branch : Polaris