[Ticket #12143] - Perform design review
Fixed writing of gzip file to use the LargeMemoryStream.  Discovered during
testing that the StudyXmlIo.WriteGzip() method was padding the data written of
the xml with null characters because of how it was grabbing the buffer from
the MemoryStream that was being used.  The usage of LargeMemoryStream also should
help memory usage.

--HG--
branch : Polaris