[Ticket #12143] - Perform design review
Reworked the writing of the Gzip to no longer write into a buffer. After some basic
performance tests, there didn't seem to be a difference in performance between
using the buffer and not using it.

Also added explicit disposal of the GZipStream, this was not being done properly
before.

--HG--
branch : Polaris