[Ticket #12801] StorageScu Component skips Group 8 Elements when Group 2 Length incorrect
Added a check to compare the actual bytes read in the group 0x0002 elements vs.
the value of tag (0002,0000).  Updated to the real bytes read.

There's a small potential flaw in this design right now.  If the transfer syntax
in group 0x0002 is explicit vr big endian, the actual first tag after group 0002
will not have been read properly, and will be byte swapped incorrectly.  This
also would have been a problem with the old code.  Decided not to address
that now, since explicit vr big endian is so infrequently used.

--HG--
branch : Polaris