[Requirement #12838] Remote Blob Store failed upload due to disposal before flush
The stack trace for the error had an exception when copying an input stream
for the RestClient into the output stream, but the output stream does not appear
to be flushed. Added a simple flush when copying the file during upload.  Not
sure if this will fix it, but want to try it out.

--HG--
branch : Polaris